🎨 ✏️ 🐛 So much bug squashing
Fix checking for inclusion by alternate sources (pf2 and 5e)

5e:
Add additional sources to optional features associated with classes;
Specifically, Artificer is from TCE, but it is included in ERLW.
For optional features related to Artificer that are also in TCE,
add ERLW as an additional source to the optional feature group.
This ensures that the group index will get pulled in.

Others:
- Fix and linkify feat prerequisites (switch to enum fields)
- Consolidate linkfy: vehicle, optional feature are with the rest
- Consolidate name decoration on the index type
- Fix race tags (slugified instead of slashified)
- ensure appendFootnotes is called across types (homebrew)
- Fix name on nodes with _copy
- Fix dice patterns, ability/savingthrow/skillCheck patterns
- Support additional formatting tags (homebrew)
- Cache result of type|name -> markdown link

Check output of the CLI driven tests for unresolved tags and unresolvable links
(not just the type-oriented tests)