Add bestiary support (#423)
* ✨ 🐉 Pf2e Bestiary

* Add ability to export creatures with the CLI tool - just the name and level for now
* Add defenses export to Pf2e creature support
* Add support for perception to pf2e creature
* Put AC note text in "notes" through tokenization

Previously only note text in the "note" key was tokenized. This makes tests fail in creature stat blocks, which use e.g.
ac:
  "notes": "{@spell mage armor}"

* Add a creature2md template file
* Add tests for the creature type
* Add `std` as an enum type rather than using the string directly.
* Add support for parsing languages for creatures.

Added a `CreatureLanguages` type as an intermediate object to hold the
language data. Not sure whether this is the desired style for these helper
classes, or whether I should be prefacing the names with `Qute` and putting
them in e.g. `QuteCreature` instead.

* Move CreatureLanguages into QuteCreature, and fix up generated
documentation.

Add support for `MarkdownDoclet` to correctly format and parse record
components. This was actually a pain to figure out, as it turns out that
these aren't accessible as an Element and instead have to be parsed through
the @param tags of the class comment.

* Add support for skills to the creature importer
* Collapse spaces in docstrings

Prevents docstring spacing from rendering weird in
markdown (such as indentation in @param tags)

* Add a ^statblock block tag to the default creature template

Also fix some import order formatting that I accidentally changed earlier.

* Add support for parsing languages for creatures.

Add support for `MarkdownDoclet` to correctly format and parse record
components. This was actually a pain to figure out, as it turns out that
these aren't accessible as an Element and instead have to be parsed through
the @param tags of the class comment.

* Use admonition block in the creature2md template