📝 Allow javadoc to use unqualified names
Currently this works by trying to call Elements.getTypeElement
on each included package, and returning the qualified name for the
retrieved TypeElement only if it's unique. That means that this will
only work for unique class names.

This falls back to the old behavior, so we should still get an error
if a link has been incorrectly qualified.