♻️ Use target as the source for dynamic text
This is a very minor change to make it easier to pull the mod
application out into another function. I've done it a separate commit to
make it easier to see and because the comment just above this line makes
me a little leery that it's intentional that this takes copyTo instead
of target - but looking at the code I can't see any reason that would
actually make a difference.